+++ app/src/main/java/com/elta/android/AppModule.kt
@@ -4,8 +4,6 @@ import android.content.Context
 import android.content.SharedPreferences
 import com.elta.android.common.di.qualifires.ComputationFacade
 import com.elta.android.common.logger.ReleaseTree
-import com.elta.android.data.features.auth.storage.LocalTokenStorage
-import com.elta.android.data.features.auth.storage.TokenStorage
 import com.elta.android.presentation.core.pm.ExceptionParser
 import com.elta.android.presentation.core.pm.SimpleExceptionParser
 import com.nullgr.core.adapter.DiffCalculator
@@ -69,8 +67,4 @@ class AppModule(private val enableLog: Boolean) {
     @Singleton
     fun provideCryptoPreferences(context: Context): CryptoPreferences =
         CryptoPreferences(context, context.getString(R.string.crypto_key_alias))
-
-    @Provides
-    @Singleton
-    fun provideTokenStorage(pref: CryptoPreferences): TokenStorage = LocalTokenStorage(pref)
 }